Board Size: Prevent the wrong size being set
When setting the board size a check is made to see if the screen is big
enough. This follows a series of checks, starting with the largest then
checking for medium and finally small. Unfortunately, the small setting
ends up being set for both the large and small size. Resolve this by
fixing the flow.