Fixed an issue with launching in windowed mode instead
The fix comes from an old script written specifically for Overworld
to launch the app in fullscreen with. I cannot understand why I did
not add that line to the main binary instead to fix this for
everyone at that time. With this change, the game should now launch
in fullscreen mode from anywhere in the OS.